Tratando de completar solicitar

+++ EzDeal/API JAVA/src/main/java/com/myorg/ezdeal/controller/ReseñaController.java
@@ -22,14 +22,14 @@ public class ReseñaController {
 
 
     @GetMapping("/{id}")
-    public List<Reseña> listarReseñasPorServicio(@PathVariable Integer id) throws Exception {
+    public List<Reseña> listarReseñasPorServicio(@PathVariable Long id) throws Exception {
         return this.reseñaService.listarReseñasPorServicio(id);
     }
 
     @PostMapping("/publicar")
     public Reseña publicarReseña(@RequestBody Reseña reseña,
-                                 @RequestParam(value="servicioId", required=true) Integer servicioId,
-                                 @RequestParam(value="clienteId", required=true) Integer clienteId) throws Exception {
+                                 @RequestParam(value="servicioId", required=true) Long servicioId,
+                                 @RequestParam(value="clienteId", required=true) Long clienteId) throws Exception {
         return this.reseñaService.publicarReseña(reseña,clienteId ,servicioId);
     }
