+++ EzDeal/API JAVA/src/main/java/com/myorg/ezdeal/controller/ServicioController.java
@@ -22,8 +22,8 @@ public class ServicioController {
 
     @RequestMapping(path="/publicar", method = RequestMethod.POST)
     public Servicio publicarServicio(@RequestBody Servicio servicio,
-                                     @RequestParam(value = "tipoServicioId" , required = true) Integer tipoServicioId,
-                                     @RequestParam(value="anuncianteId", required = true) Integer anuncianteId) throws Exception{
+                                     @RequestParam(value = "tipoServicioId" , required = true) Long tipoServicioId,
+                                     @RequestParam(value="anuncianteId", required = true) Long anuncianteId) throws Exception{
 
 
         log.info("***********************");
