+++ EzDeal/API JAVA/src/main/java/com/myorg/ezdeal/service/ReseñaService.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface ReseñaService{
 
-    Reseña publicarReseña(Reseña reseña, Integer clienteId, Integer servicioId) throws Exception;
-    List<Reseña> listarReseñasPorServicio(Integer servicioId) throws Exception;
+    Reseña publicarReseña(Reseña reseña, Long clienteId, Long servicioId) throws Exception;
+    List<Reseña> listarReseñasPorServicio(Long servicioId) throws Exception;
 }
